# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Employee Management System API",
    "description": "REST API backing the Employee Management System's web and mobile clients, covering three modules: employee & identity (employee/department records and JWT login with Admin/Manager/Employee roles), attendance (self-service daily check-in/check-out and history), and leave (leave requests and the Manager/Admin approval workflow).",
    "version": "v1"
  },
  "paths": {
...
    "/api/leave-requests": {
      "post": {
        "tags": [
          "LeaveRequests"
        ],
        "summary": "Submits a new leave request for the authenticated employee.",
        "description": "Any authenticated role. Always created for the caller's own id (from the JWT) and always\r\nstarts as `Pending` — only the decision endpoint can move it to\r\n`Approved`/`Rejected`. Returns 400 if `EndDate` is before `StartDate`.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveRequestCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveRequestCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveRequestCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "LeaveRequests"
        ],
        "summary": "Lists leave requests, optionally filtered by employee and status.",
        "description": "Role-scoped visibility: an `Employee` always sees only their own requests — the\r\nemployeeId filter is silently ignored for that role, same rule as\r\nAttendance's history endpoint. `Manager`/`Admin` may filter across everyone's\r\nrequests. status (`Pending`/`Approved`/`Rejected`)\r\napplies to every role.",
        "parameters": [
          {
            "name": "employeeId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/LeaveStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaveResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/leave-requests/{id}/decision": {
      "put": {
        "tags": [
          "LeaveRequests"
        ],
        "summary": "Approves or rejects a pending leave request.",
        "description": "Manager or Admin only. A `Manager` may only decide on their own direct reports'\r\nrequests (checked via the employee's `ManagerId`) — deciding on someone outside their\r\nteam returns 403; `Admin` can decide on anyone's. `Status` must be\r\n`Approved` or `Rejected` (400 for `Pending`). Only one decision is allowed\r\nper request — deciding on a request that isn't currently `Pending` returns 409.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveDecisionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveDecisionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveDecisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "LeaveDecisionDto": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/LeaveStatus"
          }
        },
        "additionalProperties": false
      },
      "LeaveRequestCreateDto": {
        "required": [
          "endDate",
          "leaveType",
          "startDate"
        ],
        "type": "object",
        "properties": {
          "leaveType": {
            "$ref": "#/components/schemas/LeaveType"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "reason": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LeaveResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "employeeId": {
            "type": "integer",
            "format": "int32"
          },
          "employeeName": {
            "type": "string",
            "nullable": true
          },
          "leaveType": {
            "type": "string",
            "nullable": true
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "approvedById": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "approvedByName": {
            "type": "string",
            "nullable": true
          },
          "appliedOn": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "LeaveStatus": {
        "enum": [
          "Pending",
          "Approved",
          "Rejected"
        ],
        "type": "string",
        "description": "Lifecycle state of a leave request, driven by the Manager/Admin approval workflow."
      },
      "LeaveType": {
        "enum": [
          "Sick",
          "Vacation",
          "Casual",
          "Unpaid",
          "Other"
        ],
        "type": "string",
        "description": "Category selected when an employee submits a leave request."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT bearer token. Obtain one from POST /api/auth/login, then paste just the token here (Swagger UI adds the 'Bearer ' prefix for you).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}